move H2Integrate model to top-level import
(cherry picked from commit 62a574e8a7389a6fd51037690e1035c497ece979)
