+++ docs/control/controller_demonstrations.md
@@ -16,7 +16,7 @@ kernelspec:
 ```{code-cell} ipython3
 from pathlib import Path
 from matplotlib import pyplot as plt
-from h2integrate.core.h2integrate_model import H2IntegrateModel
+from h2integrate import H2IntegrateModel
 from h2integrate import EXAMPLE_DIR
 
 ```
