+++ examples/01_onshore_steel_mn/run_onshore_steel_mn.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from h2integrate.core.h2integrate_model import H2IntegrateModel
+from h2integrate import H2IntegrateModel
 
 
 # Create a H2Integrate model
