+++ examples/02_texas_ammonia/run_texas_ammonia_plant.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from h2integrate.core.h2integrate_model import H2IntegrateModel
+from h2integrate import H2IntegrateModel
 from h2integrate.postprocess.sql_timeseries_to_csv import save_case_timeseries_as_csv
