+++ examples/03_methanol/co2_hydrogenation/run_co2h_methanol.py
@@ -1,4 +1,4 @@
-from h2integrate.core.h2integrate_model import H2IntegrateModel
+from h2integrate import H2IntegrateModel
 
 
 # Create an H2I model
