+++ examples/05_wind_h2_opt/run_wind_electrolyzer.py
@@ -1,4 +1,4 @@
-from h2integrate.core.h2integrate_model import H2IntegrateModel
+from h2integrate import H2IntegrateModel
 
 
 # Create a GreenHEART model
