+++ examples/09_co2/direct_ocean_capture/run_wind_wave_doc.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from h2integrate.core.h2integrate_model import H2IntegrateModel
+from h2integrate import H2IntegrateModel
 
 
 # Create a GreenHEART model
