+++ examples/10_electrolyzer_om/run_electrolyzer_om.py
@@ -1,4 +1,4 @@
-from h2integrate.core.h2integrate_model import H2IntegrateModel
+from h2integrate import H2IntegrateModel
 
 
 # Create a GreenHEART model
