+++ examples/12_ammonia_synloop/run_ammonia_synloop.py
@@ -2,8 +2,8 @@ from pathlib import Path
 
 import numpy as np
 
+from h2integrate import H2IntegrateModel
 from h2integrate.tools.run_cases import modify_tech_config, load_tech_config_cases
-from h2integrate.core.h2integrate_model import H2IntegrateModel
 
 
 # Create a H2Integrate model
