+++ examples/13_dispatch_for_electrolyzer/run_dispatch_for_electrolyzer.py
@@ -1,7 +1,6 @@
 from pathlib import Path
 
-from h2integrate import EXAMPLE_DIR
-from h2integrate.core.h2integrate_model import H2IntegrateModel
+from h2integrate import EXAMPLE_DIR, H2IntegrateModel
 from h2integrate.core.inputs.validation import load_tech_yaml, load_plant_yaml, load_driver_yaml
