+++ examples/14_wind_hydrogen_dispatch/hydrogen_dispatch.py
@@ -4,7 +4,7 @@ example.
 
 from pathlib import Path
 
-from h2integrate.core.h2integrate_model import H2IntegrateModel
+from h2integrate import H2IntegrateModel
 
 
 config_file = Path("./inputs/h2i_wind_to_h2_storage.yaml").resolve()
