+++ examples/15_wind_solar_electrolyzer/run_wind_solar_electrolyzer.py
@@ -1,4 +1,4 @@
-from h2integrate.core.h2integrate_model import H2IntegrateModel
+from h2integrate import H2IntegrateModel
 
 
 # Create a H2Integrate model
