+++ examples/21_iron_examples/iron_cmu/scrap_only/run_steel.py
@@ -1,4 +1,4 @@
-from h2integrate.core.h2integrate_model import H2IntegrateModel
+from h2integrate import H2IntegrateModel
 
 
 config_path = "single_site_steel.yaml"
