+++ examples/23_solar_wind_ng_demand/run_solar_wind_ng_demand.py
@@ -1,4 +1,4 @@
-from h2integrate.core.h2integrate_model import H2IntegrateModel
+from h2integrate import H2IntegrateModel
 
 
 ##################################
