+++ examples/27_site_doe_diff/run_site_doe.py
@@ -4,7 +4,7 @@ from pathlib import Path
 import pandas as pd
 import openmdao.api as om
 
-from h2integrate.core.h2integrate_model import H2IntegrateModel
+from h2integrate import H2IntegrateModel
 
 
 os.chdir(Path(__file__).parent)
