+++ examples/33_peak_load_management/run_peak_load_management.py
@@ -27,8 +27,8 @@ import pandas as pd
 import matplotlib.pyplot as plt
 from matplotlib.patches import Patch
 
+from h2integrate import H2IntegrateModel
 from h2integrate.core.utilities import build_time_series_from_plant_config
-from h2integrate.core.h2integrate_model import H2IntegrateModel
 
 
 # Create, setup, and run the H2Integrate model
