+++ h2integrate/__init__.py
@@ -6,3 +6,6 @@ ROOT_DIR = Path(__file__).resolve().parent
 EXAMPLE_DIR = ROOT_DIR.parent / "examples"
 RESOURCE_DEFAULT_DIR = ROOT_DIR.parent / "resource_files"
 H2I_LIBRARY_DIR = ROOT_DIR.parent / "library"
+
+# isort: off
+from h2integrate.core.h2integrate_model import H2IntegrateModel
