+++ h2integrate/converters/hydrogen/test/test_size_modes.py
@@ -2,8 +2,7 @@ import os
 
 import pytest
 
-from h2integrate import EXAMPLE_DIR
-from h2integrate.core.h2integrate_model import H2IntegrateModel
+from h2integrate import EXAMPLE_DIR, H2IntegrateModel
 from h2integrate.core.inputs.validation import load_tech_yaml, load_plant_yaml, load_driver_yaml
