+++ h2integrate/core/test/test_framework.py
@@ -9,8 +9,7 @@ import numpy as np
 import pytest
 
 import h2integrate.core.h2integrate_model as h2i_model_module
-from h2integrate import ROOT_DIR, EXAMPLE_DIR
-from h2integrate.core.h2integrate_model import H2IntegrateModel
+from h2integrate import ROOT_DIR, EXAMPLE_DIR, H2IntegrateModel
 from h2integrate.core.inputs.validation import load_tech_yaml, load_plant_yaml, load_driver_yaml
