+++ h2integrate/postprocess/test/test_sql_to_csv.py
@@ -13,8 +13,7 @@ import pytest
 import openmdao.api as om
 from pytest import fixture
 
-from h2integrate import EXAMPLE_DIR
-from h2integrate.core.h2integrate_model import H2IntegrateModel
+from h2integrate import EXAMPLE_DIR, H2IntegrateModel
 from h2integrate.core.inputs.validation import load_yaml, load_tech_yaml, load_driver_yaml
 from h2integrate.postprocess.sql_to_csv import summarize_case, convert_sql_to_csv_summary
