+++ h2integrate/tools/test/test_tools.py
@@ -3,9 +3,8 @@ from pathlib import Path
 
 import pytest
 
-from h2integrate import EXAMPLE_DIR
+from h2integrate import EXAMPLE_DIR, H2IntegrateModel
 from h2integrate.tools.run_cases import modify_tech_config, load_tech_config_cases
-from h2integrate.core.h2integrate_model import H2IntegrateModel
 
 
 @pytest.mark.regression
